+++ ignite/templates/ibc/placeholders.go
@@ -1,9 +1,0 @@
-package ibc
-
-//nolint:godot
-const (
-	// Placeholders IBC packets
-	PlaceholderIBCPacketModuleRecv    = "// this line is used by starport scaffolding # ibc/packet/module/recv"
-	PlaceholderIBCPacketModuleAck     = "// this line is used by starport scaffolding # ibc/packet/module/ack"
-	PlaceholderIBCPacketModuleTimeout = "// this line is used by starport scaffolding # ibc/packet/module/timeout"
-)